Tweaked background, sorted code, removed unnecessary files
