+++ Nimbus/src/components/Background.vue
@@ -2,7 +2,7 @@
   <transition name="fade" v-if="isDay" v-for="cloud in clouds" class="h-100%">
       <img :src="`/src/assets/background/cloud${rand(1, 6)}.png`" alt="" 
       :style="`animation-delay: -${rand(0, 150)};`"
-      :class="`cloud scale${rand(0, 5)} duration${rand(0, 5)} start${rand(0, 5)} position${rand(0, 5)} absolute -z-1 mix-blend-`"/>
+      :class="`cloud scale${rand(0, 5)} duration${rand(0, 5)} start${rand(0, 5)} position${rand(0, 5)} absolute -z-1 mix-blend-overlay`"/>
   </transition>
   <transition name="fade" v-if="!isDay" v-for="cloud in clouds" class="h-100%">
       <img :src="`/src/assets/background/cloud${rand(1, 6)}.png`" alt="" 
@@ -15,33 +15,35 @@
   </div>
   <img v-if="isDay && clouds < 15" :src="`/src/assets/background/sun1.png`" alt="" class="absolute top-[-10%] left-0 -z-1 scale-50"/>
 </template>
+
 <script setup lang="ts">
-import { ref, computed } from 'vue';
-import { useWeather, useIsDay } from '../composables/AppData'
+  import { ref, computed } from 'vue';
+  import { useWeather, useIsDay } from '../composables/AppData'
 
-const weatherData = useWeather()
-const currentWeather = weatherData.value.timelines.minutely[0].values
-const clouds = ref(Math.round((currentWeather.cloudCover / 10) * 2));
-const isDay = computed(() => {
-  const weather = weatherData.value
-  if (!weather || !weather.timelines?.hourly?.[0]?.time) return null
-  return useIsDay(weather.timelines.hourly[0].time)
-})
-console.log(clouds.value);
+  const weatherData = useWeather()
+  const currentWeather = weatherData.value.timelines.minutely[0].values
+  const clouds = ref(Math.round((currentWeather.cloudCover / 10) * 2));
+  const isDay = computed(() => {
+    const weather = weatherData.value
+    if (!weather || !weather.timelines?.hourly?.[0]?.time) return null
+    return useIsDay(weather.timelines.hourly[0].time)
+  })
+  console.log(clouds.value);
 
-function rand(min: number | null, max: number) {
-  let number = 0;
+  function rand(min: number | null, max: number) {
+    let number = 0;
 
-  if (min) {
-    number = Math.floor(Math.random() * max) + min;
-  } else {
-    number = Math.floor(Math.random() * max)
+    if (min) {
+      number = Math.floor(Math.random() * max) + min;
+    } else {
+      number = Math.floor(Math.random() * max)
+    }
+
+    return number;
   }
 
-  return number;
-}
-
 </script>
+
 <style scoped>
   .cloud{
     animation-name: hover; 
@@ -53,11 +55,11 @@ function rand(min: number | null, max: number) {
   .position3{top: 10%;}
   .position4{top: 15%;}
 
-  .duration0{animation-duration: 90s;}
-  .duration1{animation-duration: 100s;}
-  .duration2{animation-duration: 110s;}
-  .duration3{animation-duration: 120s;}
-  .duration4{ animation-duration: 150s;}
+  .duration0{animation-duration: 50s;}
+  .duration1{animation-duration: 60s;}
+  .duration2{animation-duration: 70s;}
+  .duration3{animation-duration: 80s;}
+  .duration4{ animation-duration: 110s;}
 
   .scale0{scale: 150%;}
   .scale1{scale: 200%;}
