+++ Nimbus/src/components/CurrentWeather.vue
@@ -1,5 +1,5 @@
 <template>
-  <div v-if="currentWeather" class="card bg-gradient-to-br from-white/90 to-gray-100/80 rounded-2xl shadow-xl p-8 w-full max-w-md mx-auto mb-6 text-black flex flex-col items-center">
+  <div v-if="currentWeather" class="card bg-gradient-to-br from-white/30 to-gray-100/30 rounded-2xl shadow-xl p-8 w-full max-w-md mx-auto mb-6 text-black flex flex-col items-center">
     <div class="flex flex-col items-center mb-4">
       <img :src="useWeatherIcon(currentWeather.weatherCode, currentWeather.time)">
       <span class="text-5xl font-extrabold tracking-tight mb-1">{{ currentWeather.temperature }}°C</span>
