--- Nimbus/src/composables/DaysData.ts
@@ -1,6 +1,0 @@
-
-
-export function useFormatDay(dateStr: string) {
-  const date = new Date(dateStr)
-  return date.toLocaleDateString('en-US', { weekday: 'long' })
-}
